+++ wp-content/plugins/event-tickets-plus/src/Tribe/Commerce/WooCommerce/Main.php
@@ -25,7 +25,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	 * Prefix used to generate the key of the transient to be associated with a different user session to avoid redirections
 	 * on users with no transient present.
 	 *
-	 * @since TBD
+	 * @since 4.7.3
 	 *
 	 * @var string
 	 */
@@ -208,7 +208,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	/**
 	 * Min required The Events Calendar version
 	 */
-	const REQUIRED_TEC_VERSION = '4.6';
+	const REQUIRED_TEC_VERSION = '4.6.20';
 
 	/**
 	 * Min required WooCommerce version
@@ -267,6 +267,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 		add_action( 'before_delete_post', array( $this, 'handle_delete_post' ) );
 		add_action( 'woocommerce_checkout_update_order_meta', array( $this, 'generate_tickets' ) );
 		add_action( 'woocommerce_order_status_changed', array( $this, 'generate_tickets' ), 12 );
+		add_action( 'woocommerce_order_status_changed', array( $this, 'reset_attendees_cache' ) );
 		add_action( 'woocommerce_email_header', array( $this, 'maybe_add_tickets_msg_to_email' ), 10, 2 );
 		add_action( 'tribe_events_tickets_metabox_edit_advanced', array( $this, 'do_metabox_advanced_options' ), 10, 2 );
 
@@ -293,11 +294,20 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 		add_action( 'woocommerce_order_action_resend_tickets_email', array( $this, 'send_tickets_email' ) );    // WC 3.2.x
 
 		add_filter( 'event_tickets_attendees_woo_checkin_stati', tribe_callback( 'tickets-plus.commerce.woo.checkin-stati', 'filter_attendee_ticket_checkin_stati' ), 10 );
+		add_action( 'wootickets_checkin', array( $this, 'purge_attendees_transient' ) );
+		add_action( 'wootickets_uncheckin', array( $this, 'purge_attendees_transient' ) );
 		add_filter( 'tribe_tickets_settings_post_types', array( $this, 'exclude_product_post_type' ) );
 
 		add_action( 'tribe_tickets_attendees_page_inside', array( $this, 'render_tabbed_view' ) );
 		add_action( 'woocommerce_check_cart_items', array( $this, 'validate_tickets' ) );
 		add_action( 'template_redirect', array( $this, 'redirect_to_cart' ) );
+
+		add_action( 'wc_after_products_starting_sales', array( $this, 'syncronize_products' ) );
+		add_action( 'wc_after_products_ending_sales', array( $this, 'syncronize_products' ) );
+
+		add_filter( 'tribe_tickets_get_ticket_max_purchase', array( $this, 'filter_ticket_max_purchase' ), 10, 2 );
+
+		tribe_singleton( 'commerce.woo.order.refunded', 'Tribe__Tickets_Plus__Commerce__WooCommerce__Orders__Refunded' );
 	}
 
 	public function register_resources() {
@@ -440,6 +450,17 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	public function admin_enqueue_styles() {
 		wp_enqueue_style( 'TribeEventsWooTickets' );
 		wp_enqueue_style( 'tribe-events-wootickets-override-style' );
+	}
+
+	/**
+	 * Where the cart form should lead the users into
+	 *
+	 * @since  4.8.1
+	 *
+	 * @return string
+	 */
+	public function get_cart_url() {
+		return wc_get_cart_url();
 	}
 
 	/**
@@ -667,6 +688,10 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 				// Iterate over all the amount of tickets purchased (for this product)
 				$quantity = (int) $item['qty'];
 
+				/** @var Tribe__Tickets__Commerce__Currency $currency */
+				$currency        = tribe( 'tickets.commerce.currency' );
+				$currency_symbol = $currency->get_currency_symbol( $product_id, true );
+
 				for ( $i = 0; $i < $quantity; $i ++ ) {
 
 					$attendee = array(
@@ -686,6 +711,8 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 						update_post_meta( $attendee_id, self::ATTENDEE_EVENT_KEY, $post_id );
 						update_post_meta( $attendee_id, self::ATTENDEE_OPTOUT_KEY, $optout );
 						update_post_meta( $attendee_id, $this->security_code, $this->generate_security_code( $order_id, $attendee_id ) );
+						update_post_meta( $attendee_id, '_paid_price', $this->get_price_value( $product_id ) );
+						update_post_meta( $attendee_id, '_price_currency_symbol', $currency_symbol );
 
 						/**
 						 * WooCommerce-specific action fired when a WooCommerce-driven attendee ticket for an event is generated
@@ -907,7 +934,9 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 				$sku = $raw_data['ticket_sku'];
 			} else {
 				$post_author = get_post_field( 'post_author', $ticket->ID );
-				$sku         = "{$ticket->ID}-{$post_author}-" . sanitize_title( $raw_data['ticket_name'] );
+				$str         = $raw_data['ticket_name'];
+				$str         = mb_strtoupper( $str, mb_detect_encoding( $str ) );
+				$sku         = "{$ticket->ID}-{$post_author}-" . str_replace( ' ', '-', $str );
 			}
 
 			update_post_meta( $ticket->ID, '_sku', $sku );
@@ -1096,19 +1125,6 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 			delete_post_meta( $ticket->ID, '_ticket_end_date' );
 		}
 
-		// Default Purchase Limit
-		if ( ! isset( $ticket->purchase_limit ) ) {
-			$ticket->purchase_limit = '';
-		}
-
-		$ticket->purchase_limit = trim( Tribe__Utils__Array::get( $raw_data, 'ticket_purchase_limit', $ticket->purchase_limit ) );
-
-		if ( '' !== $ticket->purchase_limit ) {
-			update_post_meta( $ticket->ID, '_ticket_purchase_limit', absint( $ticket->purchase_limit ) );
-		} else {
-			delete_post_meta( $ticket->ID, '_ticket_purchase_limit' );
-		}
-
 		tribe( 'tickets.version' )->update( $ticket->ID );
 
 		/**
@@ -1149,8 +1165,20 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 		}
 		$product_id = get_post_meta( $ticket_id, $this->attendee_product_key, true );
 
-		// Try to kill the actual ticket/attendee post
-		$delete = wp_delete_post( $ticket_id, true );
+		/**
+		 * Use this Filter to choose if you want to trash tickets instead
+		 * of deleting them directly
+		 *
+		 * @param bool   false
+		 * @param int    $ticket_id
+		 */
+		if ( apply_filters( 'tribe_tickets_plus_trash_ticket', true, $ticket_id ) ) {
+			// Move it to the trash
+			$delete = wp_trash_post( $ticket_id );
+		} else {
+			// Try to kill the actual ticket/attendee post
+			$delete = wp_delete_post( $ticket_id, true );
+		}
 
 		if ( is_wp_error( $delete ) ) {
 			return false;
@@ -1280,7 +1308,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 		$expired_tickets = 0;
 
 		foreach ( $tickets as $ticket ) {
-			if ( ! $ticket->date_in_range( current_time( 'timestamp' ) ) ) {
+			if ( ! $ticket->date_in_range() ) {
 				$expired_tickets++;
 			}
 		}
@@ -1302,13 +1330,12 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 			 *
 			 * add_action( 'tribe_tickets_expired_front_end_ticket_form', array( Tribe__Tickets_Plus__Attendees_List::instance(), 'render' ) );
 			 *
-			 * @since TBD
+			 * @since 4.7.3
 			 *
 			 * @param boolean $must_login
 			 * @param array $tickets
 			 */
 			do_action( 'tribe_tickets_expired_front_end_ticket_form', $must_login, $tickets );
-			return;
 		}
 
 		$global_stock_enabled = $this->uses_global_stock( $post->ID );
@@ -1392,7 +1419,6 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 		$return->provider_class   = get_class( $this );
 		$return->admin_link       = admin_url( sprintf( get_post_type_object( $product_post->post_type )->_edit_link . '&action=edit', $ticket_id ) );
 		$return->report_link      = $this->get_ticket_reports_link( null, $ticket_id );
-		$return->purchase_limit   = get_post_meta( $ticket_id, '_ticket_purchase_limit', true );
 		$return->sku              = $product->get_sku();
 		$return->show_description = $return->show_description();
 
@@ -1433,6 +1459,8 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 
 		$return->qty_sold( $qty_sold );
 		$return->qty_cancelled( $this->get_cancelled( $ticket_id ) );
+		$return->qty_refunded( $this->get_refunded( $ticket_id ) );
+
 
 		// From Event Tickets 4.4.9 onwards we can supply a callback to calculate the number of
 		// pending items per ticket on demand (since determing this is expensive and the data isn't
@@ -1447,17 +1475,6 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 			// If an earlier version of Event Tickets is activated we'll need to calculate this up front
 			$pending_totals = $this->count_order_items_by_status( $ticket_id, 'incomplete' );
 			$return->qty_pending( $pending_totals['total'] ? $pending_totals['total'] : 0 );
-		}
-
-		if ( empty( $return->purchase_limit ) && 0 !== (int) $return->purchase_limit ) {
-			/**
-			 * Filter the default purchase limit for the ticket
-			 *
-			 * @param int
-			 *
-			 * @return int
-			 */
-			$return->purchase_limit = apply_filters( 'tribe_tickets_default_purchase_limit', 0 );
 		}
 
 		/**
@@ -1753,7 +1770,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 
 			case self::ATTENDEE_OBJECT :
 
-				$attendees = $this->get_attendees_by_attendee_id( $post_id );
+				$attendees = $this->get_all_attendees_by_attendee_id( $post_id );
 
 				break;
 
@@ -1841,7 +1858,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	 *
 	 * @return array
 	 */
-	protected function get_attendees_by_attendee_id( $attendee_id ) {
+	public function get_all_attendees_by_attendee_id( $attendee_id ) {
 
 		$attendees_query = new WP_Query( array(
 			'p'         => $attendee_id,
@@ -2104,6 +2121,30 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	}
 
 	/**
+	 * Remove the Post Transients when a WooCommerce Ticket is Checked In
+	 *
+	 * @since 4.8.0
+	 *
+	 * @param  int $attendee_id
+	 * @return void
+	 */
+	public function purge_attendees_transient( $attendee_id ) {
+
+		$event_id = get_post_meta( $attendee_id, $this->attendee_event_key, true );
+		if ( ! $event_id ) {
+			return;
+		}
+
+		$current_transient = Tribe__Post_Transient::instance()->get( $event_id, Tribe__Tickets__Tickets::ATTENDEES_CACHE );
+		if ( ! $current_transient ) {
+			return;
+		}
+
+		Tribe__Post_Transient::instance()->delete( $event_id, Tribe__Tickets__Tickets::ATTENDEES_CACHE, $current_transient );
+
+	}
+
+	/**
 	 * Add the extra options in the admin's new/edit ticket metabox
 	 *
 	 * @param $post_id
@@ -2127,15 +2168,6 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 			$event_capacity = tribe_tickets_get_capacity( $post_id );
 		}
 
-		/**
-		 * Filter the default purchase limit for the ticket
-		 *
-		 * @param int
-		 *
-		 * @return int
-		 */
-		$purchase_limit = apply_filters( 'tribe_tickets_default_purchase_limit', 0 );
-
 		if ( ! empty( $ticket_id ) ) {
 			$ticket = $this->get_ticket( $post_id, $ticket_id );
 			$is_correct_provider = tribe( 'tickets.handler' )->is_correct_provider( $ticket_id, $this );
@@ -2145,10 +2177,6 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 				$capacity            = tribe_tickets_get_capacity( $ticket->ID );
 				$global_stock_mode   = ( method_exists( $ticket, 'global_stock_mode' ) ) ? $ticket->global_stock_mode() : '';
 				$global_stock_cap    = ( method_exists( $ticket, 'global_stock_cap' ) ) ? $ticket->global_stock_cap() : 0;
-
-				if ( metadata_exists( 'post', $ticket->ID, '_ticket_purchase_limit' ) ) {
-					$purchase_limit = get_post_meta( $ticket->ID, '_ticket_purchase_limit', true );
-				}
 			}
 		}
 
@@ -2500,6 +2528,20 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 		return $cancelled->get_count();
 	}
 
+	/*
+	 * Get the number of refunded orders for a ticket.
+	 * Works with partial and full refunds.
+	 *
+	 * @since 4.7.3
+	 *
+	 * @param int $ticket_id Ticket ID
+	 *
+	 * @return int
+	 */
+	private function get_refunded( $ticket_id ) {
+		return tribe( 'commerce.woo.order.refunded' )->get_count( $ticket_id );
+	}
+
 	/**
 	 * @param $order_id
 	 */
@@ -2514,6 +2556,30 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 		 * @param int $order_id The order post ID for the ticket.
 		 */
 		do_action( 'event_tickets_woo_complete_order', $order_id );
+	}
+
+	/**
+	 * Filter the Quantity of max purchase for WooCommerce
+	 *
+	 * @since  4.8.1
+	 *
+	 * @param int                           $available Max Purchase number
+	 * @param Tribe__Tickets__Ticket_Object $ticket    Ticket Object
+	 *
+	 * @return int
+	 */
+	public function filter_ticket_max_purchase( $available, $ticket ) {
+		if ( class_exists( 'WC_Product_Simple' ) ) {
+			$product = new WC_Product_Simple( $ticket->ID );
+		} else {
+			$product = new WC_Product( $ticket->ID );
+		}
+
+		$stock        = $ticket->stock();
+		$max_quantity = $product->backorders_allowed() ? '' : $stock;
+		$max_quantity = $product->is_sold_individually() ? 1 : $max_quantity;
+
+		return $max_quantity;
 	}
 
 	/*
@@ -2659,7 +2725,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	 * Instead of removing the product from the cart we send a notice and avoid to checkout so the user knows exactly why can
 	 * move forward and he needs to take an action before doing so.
 	 *
-	 * @since TBD
+	 * @since 4.7.3
 	 *
 	 * @return bool
 	 */
@@ -2678,7 +2744,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 				continue;
 			}
 
-			if ( ! $ticket_type->date_in_range( current_time( 'timestamp' ) ) ) {
+			if ( ! $ticket_type->date_in_range() ) {
 
 				$message = sprintf(
 					__( 'The ticket: %1$s, in your cart is no longer available or valid. You need to remove it from your cart in order to continue.', 'event-tickets-plus' ),
@@ -2700,7 +2766,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	 *
 	 * @see https://en.wikipedia.org/wiki/Post/Redirect/Get
 	 *
-	 * @since  TBD
+	 * @since  4.7.3
 	 */
 	public function redirect_to_cart() {
 		$cart_url = get_transient( $this->get_cart_transient_key() );
@@ -2711,7 +2777,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 			 * Filter to allow the change the URL where the users are redirected by default uses the wc_get_cart_url()
 			 * value.
 			 *
-			 * @since TBD
+			 * @since 4.7.3
 			 *
 			 * @param string $location
 			 */
@@ -2727,7 +2793,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	/**
 	 * Get the key used to store the cart transient URL.
 	 *
-	 * @since TBD
+	 * @since 4.7.3
 	 *
 	 * @return string
 	 */
@@ -2738,7 +2804,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	/**
 	 * Generates as hash based on the user session, user cart or user ID
 	 *
-	 * @since TBD
+	 * @since 4.7.3
 	 *
 	 * @return string
 	 */
@@ -2758,7 +2824,7 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 	/**
 	 * Get the default Currency selected for Woo
 	 *
-	 * @since TBD
+	 * @since 4.7.3
 	 *
 	 * @return string
 	 */
@@ -2766,4 +2832,61 @@ class Tribe__Tickets_Plus__Commerce__WooCommerce__Main extends Tribe__Tickets_Pl
 		return function_exists( 'get_woocommerce_currency' ) ? get_woocommerce_currency() : parent::get_currency();
 	}
 
+	/**
+	 * Clean the attendees cache every time an order changes its
+	 * status, so the changes are reflected instantly.
+	 *
+	 * @since 4.7.3
+	 *
+	 * @param int $order_id
+	 */
+	public function reset_attendees_cache( $order_id ) {
+
+		// Get the items purchased in this order
+		$order       = new WC_Order( $order_id );
+		$order_items = $order->get_items();
+
+		// Bail if the order is empty
+		if ( empty( $order_items ) ) {
+			return;
+		}
+
+		// Iterate over each product
+		foreach ( (array) $order_items as $item_id => $item ) {
+			$product_id = isset( $item['product_id'] ) ? $item['product_id'] : $item['id'];
+			// Get the event this tickets is for
+			$post_id = get_post_meta( $product_id, $this->event_key, true );
+
+			if ( ! empty( $post_id ) ) {
+				// Delete the attendees cache for that event
+				tribe( 'post-transient' )->delete( $post_id, Tribe__Tickets__Tickets::ATTENDEES_CACHE );
+			}
+		}
+	}
+
+	/**
+	 * Sincronize the Event cost from an array of
+	 * product IDs
+	 *
+	 * @since 4.7.3
+	 *
+	 * @param array $product_ids
+	 */
+	public function syncronize_products( $product_ids ) {
+
+		if ( $product_ids ) {
+
+			foreach ( $product_ids as $product_id ) {
+				$event = $this->get_event_for_ticket( $product_id );
+
+				// This product is not connected with an event
+				if ( ! $event ) {
+					continue;
+				}
+
+				// Trigger an update
+				Tribe__Events__API::update_event_cost( $event->ID );
+			}
+		}
+	}
 }
